Multiple Fixes.

+++ src/test_simplealgebra/TestBurgersB.java
@@ -1530,7 +1530,7 @@ public class TestBurgersB extends TestCase {
 		for( int tcnt = 0 ; tcnt < 2 ; tcnt++ )
 		{
 			System.out.println( "Initial - " + tcnt );
-			for( int acnt = 0 ; acnt < NUM_X_ITER ; acnt++ )
+			for( long acnt = 0 ; acnt < NUM_X_ITER ; acnt++ )
 			{
 				atm2 = System.currentTimeMillis();
 				if( atm2 - atm >= 1000 )
@@ -1539,8 +1539,8 @@ public class TestBurgersB extends TestCase {
 					atm = atm2;
 				}
 				
-				int ac = acnt;
-				final int x = ac % NUM_X_ITER;
+				long ac = acnt;
+				final int x = (int)( ac % NUM_X_ITER );
 				final double dx = ( x - HALF_X ) / RAD_X;
 				if( dx * dx < 1.0 )
 				{
