+++ src/test_simplealgebra/TestSchrodingerA.java
@@ -2043,7 +2043,7 @@ public class TestSchrodingerA extends TestCase {
 		for( int tcnt = 0 ; tcnt < 2 ; tcnt++ )
 		{
 			System.out.println( "Initial - " + tcnt );
-			for( int acnt = 0 ; acnt < NUM_X_ITER * NUM_Y_ITER * NUM_Z_ITER ; acnt++ )
+			for( long acnt = 0 ; acnt < ( (long) NUM_X_ITER ) * NUM_Y_ITER * NUM_Z_ITER ; acnt++ )
 			{
 				atm2 = System.currentTimeMillis();
 				if( atm2 - atm >= 1000 )
@@ -2052,12 +2052,12 @@ public class TestSchrodingerA extends TestCase {
 					atm = atm2;
 				}
 				
-				int ac = acnt;
-				final int z = ac % NUM_Z_ITER;
+				long ac = acnt;
+				final int z = (int)( ac % NUM_Z_ITER );
 				ac = ac / NUM_Z_ITER;
-				final int y = ac % NUM_Y_ITER;
+				final int y = (int)( ac % NUM_Y_ITER );
 				ac = ac / NUM_Y_ITER;
-				final int x = ac % NUM_X_ITER;
+				final int x = (int)( ac % NUM_X_ITER );
 				final double dx = ( x - HALF_X ) / RAD_X;
 				final double dy = ( y - HALF_Y ) / RAD_Y;
 				final double dz = ( z - HALF_Z ) / RAD_Z;
