+++ src/test_simplealgebra/TestStelemC.java
@@ -1697,7 +1697,7 @@ public class TestStelemC extends TestCase {
 		for( int tcnt = 0 ; tcnt < 2 ; tcnt++ )
 		{
 			System.out.println( "Initial - " + tcnt );
-			for( int acnt = 0 ; acnt < NUM_X_ITER * NUM_Y_ITER ; acnt++ )
+			for( long acnt = 0 ; acnt < ( (long) NUM_X_ITER ) * NUM_Y_ITER ; acnt++ )
 			{
 				atm2 = System.currentTimeMillis();
 				if( atm2 - atm >= 1000 )
@@ -1706,10 +1706,10 @@ public class TestStelemC extends TestCase {
 					atm = atm2;
 				}
 				
-				int ac = acnt;
-				final int y = ac % NUM_Y_ITER;
+				long ac = acnt;
+				final int y = (int)( ac % NUM_Y_ITER );
 				ac = ac / NUM_Y_ITER;
-				final int x = ac % NUM_X_ITER;
+				final int x = (int)( ac % NUM_X_ITER );
 				final double dx = ( x - HALF_X ) / RAD_X;
 				final double dy = ( y - HALF_Y ) / RAD_Y;
 				if( dx * dx + dy * dy < 1.0 )
